// packages/ContentAutotagging/src/Engine/generic/content.types.d.ts
export declare class ContentItemParams {
    contentSourceID: string;
    name: string;
    description?: string;
    ContentTypeID: string;
    ContentSourceTypeID: string;
    ContentFileTypeID: string;
    URL: string;
    AIModelID?: string;
    minTags?: number;
    maxTags?: number;
}
export declare class ContentSourceParams {
    contentSourceID: string;
    name?: string;
    description?: string;
    ContentTypeID: string;
    ContentSourceTypeID: string;
    ContentFileTypeID: string;
    URL: string;
    AIModelID?: string;
    minTags?: number;
    maxTags?: number;
}
export declare class ContentSourceTypeParams {
    contentSourceID: string;
    contentSourceTypeID: string;
    name: string;
    value: string;
    type: string;
}
//# sourceMappingURL=content.types.d.ts.map